Extract Pip log analysis bases.
This is part of a series of a refactors needed to remove dependency
cycles in support of #1696.
